wip: autocomplete comp
