+++ src/components/molecules/Settings/SelectField/index.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from "react";
 
 import Icon from "@reearth/components/atoms/Icon";
 import Select, { Props as SelectProps } from "@reearth/components/atoms/Select";
-import Option from "@reearth/components/atoms/SelectOption";
+import { Option } from "@reearth/components/atoms/SelectOption";
 import Text from "@reearth/components/atoms/Text";
 import { styled } from "@reearth/theme";
